+++ shaders/pass_through.frag
@@ -10,7 +10,7 @@ uniform sampler3D levelSetTexture;
 uniform sampler3D quantityTexture;
 
 uniform float timeStep; // in microseconds
-uniform int zSlice;
+uniform float zSlice;
 
 const float step = 1.0f/gridSize;
